fix(ReviewDTO): 리뷰 DTO를 완벽하게 구성

+++ src/main/java/jeje/work/aeatbe/dto/review/ReviewDTO.java
@@ -6,13 +6,14 @@ import jeje.work.aeatbe.dto.user.UserDTO;
 import lombok.Builder;
 
 @Builder
-public record ReviewDTO(Long id,
-                        int rate,
-                        String content,
-                        UserDTO user,
-                        Long productId,
+public record ReviewDTO(
+        Long id,
+        Long rate,
+        String content,
+        UserDTO user,
+        Long productId,
 
-                        @Nullable
-                        String productImgUrl) {
+        @Nullable
+        String productImgUrl) {
 
 }
